Fixed auth bugs and cleaned up code
