+++ client/src/firebase/auth.ts
@@ -9,6 +9,7 @@ import {
 import { deleteDoc, doc, getDoc, setDoc } from 'firebase/firestore';
 import { auth, db } from './firebase';
 
+// Function used to retreive user status (signed in or not)
 export const getCurrentlySignedInUserHandler = () => {
   const user = auth.currentUser;
 
@@ -19,6 +20,7 @@ export const getCurrentlySignedInUserHandler = () => {
   }
 };
 
+// Function used to call the Firebase Signup function
 export const signUpHandler = (
   email: string,
   password: string,
@@ -30,6 +32,7 @@ export const signUpHandler = (
     .then((userCredential) => {
       const user = userCredential.user;
 
+      // Calls a DB call to create a record for the user who just signed up
       setDoc(
         doc(db, 'users', user.uid),
         {
@@ -56,6 +59,7 @@ export const signUpHandler = (
     });
 };
 
+// Function used to call the Firebase login function
 export const loginHandler = (
   email: string,
   password: string,
@@ -65,6 +69,7 @@ export const loginHandler = (
     .then((userCredential) => {
       const user = userCredential.user;
 
+      // Calls the DB to update user signedIn status in DB
       setDoc(
         doc(db, 'users', user.uid),
         {
@@ -83,6 +88,7 @@ export const loginHandler = (
     });
 };
 
+// Function used to call the Firebase Google authentication function (Used for both signups and logins)
 export const googleLoginHandler = (onSuccess: () => void) => {
   const provider = new GoogleAuthProvider();
   return signInWithPopup(auth, provider)
@@ -93,22 +99,33 @@ export const googleLoginHandler = (onSuccess: () => void) => {
       const userRef = doc(db, 'users', user.uid);
       const userDoc = await getDoc(userRef);
 
-      // Determine the value of uploadedFiles based on user existence
+      // Determine the boolean value of uploadedFiles based on user existence
       const uploadedFiles = userDoc.exists()
         ? userDoc.data().uploadedFiles
         : false;
 
-      setDoc(
-        doc(db, 'users', user.uid),
-        {
-          uid: user.uid,
-          displayName: user.displayName,
-          email: user.email,
-          signedIn: true,
-          uploadedFiles: uploadedFiles,
-        },
-        { merge: true }
-      );
+      userDoc.exists()
+        ? // Case for handling if user already exists and user is just signing in using google auth
+          setDoc(
+            doc(db, 'users', user.uid),
+            {
+              signedIn: true,
+            },
+            { merge: true }
+          )
+        : // Case for handling if user doesn't exists and user is signing up with google auth
+          setDoc(
+            doc(db, 'users', user.uid),
+            {
+              uid: user.uid,
+              displayName: user.displayName,
+              email: user.email,
+              signedIn: true,
+              uploadedFiles: uploadedFiles,
+            },
+            { merge: true }
+          );
+
       return user;
     })
     .then(() => {
@@ -121,12 +138,14 @@ export const googleLoginHandler = (onSuccess: () => void) => {
     });
 };
 
+// Function used to call the Firebase delete user function
 export const deleteUserHandler = () => {
   const user = auth.currentUser;
 
   if (user) {
     return deleteUser(user)
       .then(() => {
+        // Delete DB record of user
         deleteDoc(doc(db, 'users', user.uid));
         console.log(`User ${user.uid} deleted`);
       })
@@ -140,6 +159,7 @@ export const deleteUserHandler = () => {
   }
 };
 
+// Function used to call the Firebase logout function
 export const userLogoutHandler = () => {
   const user = auth.currentUser;
 
@@ -148,6 +168,7 @@ export const userLogoutHandler = () => {
       .then(() => {
         console.log(`User ${user.uid} signed out`);
 
+        // Updating DB record to indicate that user is signed out
         setDoc(
           doc(db, 'users', user.uid),
           {
